// src/utils/types.ts
export interface RegistroPresupuestario {
    TIPO_CEDULA:   number;
    PERIODO:       number;
    ANIO:          number;
    COD_PROG:      string;
    PROGRAMA:      string;
    COD_PROY:      string;
    PROYECTO:      string;
    COD_ACTIVIDAD: string;
    ACTIVIDAD:     string;
    COD_OBRA:      string;
    OBRA:          string;
    COD_TAREA:     string;
    TAREA:         string;
    COD_FUENTE:    number;
    PARTIDA:       number;
    ASIGNACION:    number;
    REFORMAS:      number;
    CODIFICADO:    number;
    CERTIFICACION: number;
    COMPROMISOS:   number;
    DEVENGADO:     number;
    PAGADO:        number;
}